# Implement CSV direct to parquet