--- a/lilac/sources/csv_source.py
+++ b/lilac/sources/csv_source.py
@@ -1,13 +1,16 @@
 """CSV source."""
-from typing import ClassVar, Iterable, Optional, cast
+import os
+from typing import ClassVar, Optional, cast
 
 import duckdb
 import pyarrow as pa
 from pydantic import Field
 from typing_extensions import override
 
-from ..schema import Item, arrow_schema_to_schema
-from ..source import Source, SourceSchema
+from ..data import dataset_utils
+from ..schema import PARQUET_FILENAME_PREFIX, ROWID, Schema, arrow_schema_to_schema
+from ..source import Source, SourceManifest, SourceSchema
+from ..tasks import TaskStepId
 from ..utils import download_http_files
 from .duckdb_utils import convert_path_to_duckdb, duckdb_setup
 
@@ -82,13 +85,17 @@ class CSVSource(Source):
     return self._source_schema
 
   @override
-  def yield_items(self) -> Iterable[Item]:
-    """Process the source."""
-    if not self._reader or not self._con:
-      raise RuntimeError('CSV source is not initialized.')
+  def load_to_parquet(self, output_dir: str, task_step_id: Optional[TaskStepId]) -> SourceManifest:
+    del task_step_id
 
-    for batch in self._reader:
-      yield from batch.to_pylist()
+    assert self._con, 'setup() must be called first.'
 
-    self._reader.close()
-    self._con.close()
+    out_filename = dataset_utils.get_parquet_filename(PARQUET_FILENAME_PREFIX, 0, 1)
+    filepath = os.path.join(output_dir, out_filename)
+    os.makedirs(output_dir, exist_ok=True)
+
+    self._con.sql(
+      f"""SELECT replace(CAST(uuid() AS VARCHAR), ' - ', ' ') AS {ROWID}, * FROM t"""
+    ).write_parquet(filepath, compression='zstd')
+    schema = Schema(fields=self.source_schema().fields.copy())
+    return SourceManifest(files=[out_filename], data_schema=schema, source=self)